Added a new version of F2C. Now C arpack works on 64-bit machines, need to test on 32 bit ones.

+++ include/igraph_arpack.h
@@ -228,36 +228,36 @@ __BEGIN_DECLS
 typedef struct igraph_arpack_options_t {
   /* INPUT */
   char bmat[1];			/* I-standard problem, G-generalized */
-  long int n; 			/* Dimension of the eigenproblem */
+  int n; 			/* Dimension of the eigenproblem */
   char which[2];		/* LA, SA, LM, SM, BE */
-  long int nev;                 /* Number of eigenvalues to be computed */
+  int nev;                 /* Number of eigenvalues to be computed */
   igraph_real_t tol;		/* Stopping criterion */
-  long int ncv;			/* Number of columns in V */
-  long int ldv;			/* Leading dimension of V */
-  long int ishift;		/* 0-reverse comm., 1-exact with tridiagonal */
-  long int mxiter;              /* Maximum number of update iterations to take */
-  long int nb;			/* Block size on the recurrence, only 1 works */
-  long int mode;		/* The kind of problem to be solved (1-5)
+  int ncv;			/* Number of columns in V */
+  int ldv;			/* Leading dimension of V */
+  int ishift;		/* 0-reverse comm., 1-exact with tridiagonal */
+  int mxiter;              /* Maximum number of update iterations to take */
+  int nb;			/* Block size on the recurrence, only 1 works */
+  int mode;		/* The kind of problem to be solved (1-5)
 				   1: A*x=l*x, A symmetric
 				   2: A*x=l*M*x, A symm. M pos. def.
 				   3: K*x = l*M*x, K symm., M pos. semidef.
 				   4: K*x = l*KG*x, K s. pos. semidef. KG s. indef.
 				   5: A*x = l*M*x, A symm., M symm. pos. semidef. */
-  long int start;		/* 0: random, 1: use the supplied vector */
-  long int lworkl;		/* Size of temporary storage, default is fine */
+  int start;		/* 0: random, 1: use the supplied vector */
+  int lworkl;		/* Size of temporary storage, default is fine */
   igraph_real_t sigma;          /* The shift for modes 3,4,5 */
   igraph_real_t sigmai;		/* The imaginary part of shift for rnsolve */
   /* OUTPUT */
-  long int info;		/* What happened, see docs */
-  long int ierr;		/* What happened  in the dseupd call */
-  long int noiter;		/* The number of iterations taken */
-  long int nconv;
-  long int numop;		/* Number of OP*x operations */
-  long int numopb;		/* Number of B*x operations if BMAT='G' */
-  long int numreo;		/* Number of steps of re-orthogonalizations */
+  int info;		/* What happened, see docs */
+  int ierr;		/* What happened  in the dseupd call */
+  int noiter;		/* The number of iterations taken */
+  int nconv;
+  int numop;		/* Number of OP*x operations */
+  int numopb;		/* Number of B*x operations if BMAT='G' */
+  int numreo;		/* Number of steps of re-orthogonalizations */
   /* INTERNAL */
-  long int iparam[11];
-  long int ipntr[14];
+  int iparam[11];
+  int ipntr[14];
 } igraph_arpack_options_t;
 
 /**
@@ -282,14 +282,14 @@ typedef struct igraph_arpack_options_t {
  */
 
 typedef struct igraph_arpack_storage_t {
-  long int maxn, maxncv, maxldv;
+  int maxn, maxncv, maxldv;
   igraph_real_t *v;
   igraph_real_t *workl;
   igraph_real_t *workd;
   igraph_real_t *d;
   igraph_real_t *resid;
   igraph_real_t *ax;
-  long int *select;
+  int *select;
   igraph_real_t *di;		/* These two only for non-symmetric problems */
   igraph_real_t *workev;
 } igraph_arpack_storage_t;
